Find the businesses that are common to both datasets, that is, the name and address must match
As the business names and addresses don't align perfectly between these datasets, the algorithm will find approximate matches. After running, it will produce a list of triplets consisting of the entity_id from the left dataset, the business_id
from the right dataset, and a confidence score > 0.8 (max is 1 for perfect match).